Fully Functional Announcements/Lectures/Assignments
-Rescaffolded Teacher Portal into 3 subfolders. (Announcement, Assignments, Lectures)

-Added Add, View, Edit, Delete for Announcement ,Assignments, Lectures)

-Added Sessions for all 3, and some validations.

-Also redid Online Application again in preparation for tom's meeting with sir Mon. (If Sir approves of the CR for the DB we will keep this ver.)

Note: This module is fully working now, yes. But it still is not done, we still do not have the complete ver where it binds to subjects/sections/schedules.

Also take note: Please read the document I uploaded in the group for it contains the DB Changes which are heavily needed for this to work.